refactor: `functions` module
