+++ pyproject.toml
@@ -135,7 +135,7 @@ exclude = []
 "src/_gettsim/benefits/kinderzuschl.py" = ["ARG001"]
 "src/_gettsim/benefits/unterhalt.py" = ["E501"]
 "src/_gettsim/benefits/wohngeld.py" = ["ARG001"]
-"src/_gettsim/functions.py" = ["F401"]
+"src/_gettsim/functions/all_functions.py" = ["F401"]
 "src/_gettsim/gettsim_typing.py" = ["PGH", "PLR", "SIM114"]
 "src/_gettsim/piecewise_functions.py" = ["PLR"]
 "src/_gettsim/policy_environment.py" = ["S506", "PLR"]
