+++ src/_gettsim/policy_environment.py
@@ -12,13 +12,13 @@ import yaml
 
 import _gettsim.functions  # Execute all decorators # noqa: F401
 from _gettsim.config import INTERNAL_PARAMS_GROUPS, RESOURCE_DIR
-from _gettsim.functions_loader import load_functions_for_date
+from _gettsim.functions.loader import load_functions_for_date
+from _gettsim.functions.policy_function import PolicyFunction
 from _gettsim.piecewise_functions import (
     check_thresholds,
     get_piecewise_parameters,
     piecewise_polynomial,
 )
-from _gettsim.policy_function import PolicyFunction
 
 if TYPE_CHECKING:
     from collections.abc import Callable
