+++ src/_gettsim_tests/test_functions_loader.py
@@ -5,9 +5,8 @@ from typing import TYPE_CHECKING
 
 import numpy
 import pytest
-
 from _gettsim.config import RESOURCE_DIR
-from _gettsim.functions_preprocessor import (
+from _gettsim.functions.preprocessor import (
     _create_derived_functions,
     _load_functions,
     _vectorize_func,
