+++ src/_gettsim_tests/test_interface.py
@@ -3,8 +3,8 @@ import warnings
 import numpy
 import pandas as pd
 import pytest
-
 from _gettsim.config import FOREIGN_KEYS
+from _gettsim.functions.policy_function import PolicyFunction
 from _gettsim.gettsim_typing import convert_series_to_internal_type
 from _gettsim.groupings import bg_id_numpy, wthh_id_numpy
 from _gettsim.interface import (
@@ -16,7 +16,6 @@ from _gettsim.interface import (
     compute_taxes_and_transfers,
 )
 from _gettsim.policy_environment import PolicyEnvironment
-from _gettsim.policy_function import PolicyFunction
 from _gettsim.shared import policy_info
 from gettsim import FunctionsAndColumnsOverlapWarning
