+++ src/_gettsim_tests/test_vectorization.py
@@ -3,16 +3,13 @@ import string
 
 import numpy
 import pytest
-
 from _gettsim.config import PATHS_TO_INTERNAL_FUNCTIONS, USE_JAX
 
 if USE_JAX:
     import jax.numpy
-from numpy.testing import assert_array_equal
-
-from _gettsim.functions_preprocessor import _load_functions
+from _gettsim.functions.preprocessor import _load_functions
 from _gettsim.transfers.elterngeld import (
-    elterngeld_anspruchsbedingungen_erfüllt,  # noqa: PLC2403
+    elterngeld_anspruchsbedingungen_erfüllt,
     elterngeld_geschwisterbonus_m,
 )
 from _gettsim.transfers.grundrente import grundr_bew_zeiten_avg_entgeltp
@@ -21,6 +18,7 @@ from _gettsim.vectorization import (
     make_vectorizable,
     make_vectorizable_source,
 )
+from numpy.testing import assert_array_equal
 
 # ======================================================================================
 # Backend
